Use crude approximation table for conversion to polar coordinates

+++ SourceAFIS/Matching/EdgeInfo.cs
@@ -6,7 +6,7 @@ namespace SourceAFIS.Matching
 {
     public struct EdgeInfo
     {
-        public float Length;
+        public short Length;
         public byte ReferenceAngle;
         public byte NeighborAngle;
     }
